+++ frontend/src/components/main/chart/chart.js
@@ -1,20 +1,8 @@
-import { daysOfWeek } from '../../calendar/constants';
+import { getCurrentWeek } from './currentWeek';
 
 const axis_x = [0, 10, 20, 30, 40];
 
 const makeChart = () => {
-    function getCurrentWeek() {
-        const Today = new Date();
-        const LastDate = Today.getDate();
-        const day = Today.getDay();
-        const result = [];
-        for (let i = 0; i < 6; i++) {
-            const text = LastDate + i + '(' + daysOfWeek[day + i] + ')';
-            result.push(text);
-        }
-        return result;
-    }
-
     const chart =
         /*html*/
         `
