+++ frontend/src/main.js
@@ -1,5 +1,5 @@
 import setHomeEvent from './scripts/home.js';
-import setMainEvent from './scripts/main.js';
+import { setMainEvent } from './scripts/main.js';
 import { setCalendarEvent } from './components/calendar/calendar.js';
 
 const eventdelegator = (root) => {
@@ -14,7 +14,7 @@ const eventdelegator = (root) => {
         case 'track':
             break;
         case 'attend':
-            setCalendarEvent
+            setCalendarEvent();
             break;
     }
 };
